Navbar + Set up
Finished Navbar component + reset the front-end directory structure to allow for new components to be added in (since old front-end iteration was just for demonstration). The navbar component isn't "fully" finished since the rest of the components need to be made before the actual navigation functionality can work. However, i've left an instance of it in App.js just to see and mess around with it. Feel free to delete it if anyone else needs to play around with whatever component they're making.

Side Note: The implementation of the color changes on the buttons was done in a very clunky and inelegant way. I'll change this when i figure out a better way.