Added page change logic to Navbar component and completed initial testing of page change
Coded in logic that will allow the user to click buttons on the Navbar to render different pages. The page components are contained in the pages folder. Edit components in this folder to change any given page. There is probably a better way to do this than the way I've implemented but this way will definitely work and is relatively easy to work with.